requirements.txt
# Core Data Science Libraries
numpy==1.24.3
pandas==2.0.3
scipy==1.11.1

# Machine Learning
scikit-learn==1.3.0
tensorflow==2.13.0
torch==2.0.1
keras==2.13.1
xgboost==1.7.6
lightgbm==4.0.0

# MLOps
mlflow==2.7.1
optuna==3.3.0

# Data Processing
pyspark==3.4.1
pyarrow==12.0.1
fastparquet==2023.7.0

# Kafka
kafka-python==2.0.2
confluent-kafka==2.3.0

# APIs
fastapi==0.103.1
uvicorn[standard]==0.23.2
flask==2.3.3
pydantic==2.4.2
requests==2.31.0

# Database Connectors
pymongo==4.5.0
cassandra-driver==3.28.0
redis==5.0.0
sqlalchemy==2.0.20
psycopg2-binary==2.9.9
pymysql==1.1.0

# Cloud Storage
boto3==1.28.25
azure-storage-blob==12.19.0
google-cloud-storage==2.10.0

# Data Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.17.0
dash==2.13.0

# Airflow
apache-airflow==2.7.1
apache-airflow-providers-apache-spark==4.3.0
apache-airflow-providers-docker==3.7.0

# Monitoring & Logging
prometheus-client==0.18.0
python-json-logger==2.0.7

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
click==8.1.7
tqdm==4.66.1
joblib==1.3.2

# Testing
pytest==7.4.2
pytest-cov==4.1.0
pytest-mock==3.11.1

# Code Quality
black==23.9.1
flake8==6.1.0
mypy==1.5.1
pylint==2.17.7

